+++ config/routes.rb
@@ -4,7 +4,12 @@ Rails.application.routes.draw do
     sessions: "users/sessions",
   }
 
-  root to: "posts#index"
-  resources :posts
+  root to: "api/v1/posts#index"
+  namespace 'api' do
+    namespace 'v1' do
+      resources :posts
+    end
+  end
+
   # For details on the DSL available within this file, see https://guides.rubyonrails.org/routing.html
 end
